refac: use abstract class to reuse validate logic
