+++ src/test/java/nextstep/subway/unit/PathFinderTest.java
@@ -8,6 +8,7 @@ import nextstep.subway.applicaion.dto.PathRequest;
 import nextstep.subway.applicaion.dto.PathResponse;
 import nextstep.subway.domain.Line;
 import nextstep.subway.domain.PathFinder;
+import nextstep.subway.domain.Station;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -20,36 +21,24 @@ public class PathFinderTest {
 
         @BeforeEach
         public void setUp() {
+            final Station station1 = new Station(1L, "강남역");
+            final Station station2 = new Station(2L, "역삼역");
+            final Line line = new Line("2호선", "green", station1, station2, 10);
+            lines = List.of(line);
+
             pathRequest = new PathRequest(1L, 2L);
-            lines = new ArrayList<>();
             pathFinder = new PathFinderMock();
         }
 
         @Test
         public void 경로_조회_기능() {
-            PathResponse response = pathFinder.findPath(pathRequest, lines);
-
-            assertThat(response).hasOnlyFields("distance", "stations");
+            assertThat(pathFinder.findPath(pathRequest, lines)).hasOnlyFields("distance", "stations");
         }
-
-        @Test
-        public void 출발지와_도착지가_같으면_익셉션을_던진다() {
-            pathFinder.validateRequest(pathRequest, lines);
-        }
-
 }
 
-class PathFinderMock implements PathFinder {
-
+class PathFinderMock extends PathFinder {
     @Override
-    public PathResponse findPath(PathRequest pathRequest, List<Line> lines) {
+    protected PathResponse getPath(PathRequest pathRequest, List<Line> lines) {
         return new PathResponse();
-    }
-
-    @Override
-    public void validateRequest(PathRequest request, List<Line> lines) {
-        if (request.getSource().equals(request.getTarget())) {
-            throw new IllegalArgumentException("출발역과 도착역이 같습니다.");
-        }
     }
 }
